1. Changes to route for weather-forecast

+++ src/env.py
@@ -5,6 +5,6 @@ from dotenv import load_dotenv
 load_dotenv()
 PORT = int(os.getenv('PORT'))
 API_KEY = os.getenv('API_KEY')
-LAT = int(os.getenv('LAT'))
-LON = int(os.getenv('LON'))
+LAT = float(os.getenv('LAT'))
+LON = float(os.getenv('LON'))
 UNIT = os.getenv('UNIT')
